Fix/base

+++ src/recovery/log_manager.h
@@ -52,7 +52,7 @@ public:
         printf("log_type_: %s\n", LogTypeStr[log_type_].c_str());
         printf("lsn: %d\n", lsn_);
         printf("log_tot_len: %d\n", log_tot_len_);
-        printf("log_tid: %d\n", log_tid_);
+        printf("log_tid: %lld\n", static_cast<long long>(log_tid_));
         printf("prev_lsn: %d\n", prev_lsn_);
     }
 };
